% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPSolver.R
\name{LPSolver}
\alias{LPSolver}
\title{Solver to Linear problem}
\usage{
LPSolver()
}
\value{
Start a shiny apps to add the problem and solve it.
}
\description{
This function helps to solve a linear problem
}
